forencics_site/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forencics"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "click >= 8.1",
  "flask >= 2.3",
  "omegaconf >= 2.3",
  "pyyaml >= 6.0",
]

[tool.hatch.version]
path = "forencics/version.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
server = "python -m forencics.server {args}"
